% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_wrapper.R
\name{estGip_stan}
\alias{estGip_stan}
\title{Bayesian GIRT model fitting}
\usage{
estGip_stan(y, infer = "MCMC", ...)
}
\arguments{
\item{y}{integer matrix contains only item response. Do NOT contain subject ID and group index.}

\item{infer}{a method of inference. "MCMC" or "VB"}

\item{...}{Arguments passe to \code{\link[rstan]{sampling}}}
}
\value{
stanfit class file returned by \code{\link{rstan::sampling}} or \code{\link[rstan]{vb}}.
}
\description{
Bayesian GIRT model fitting
}
