% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_wrapper.R
\name{estgrm_stan}
\alias{estgrm_stan}
\title{Bayesian Graded Response modelng fitting}
\usage{
estgrm_stan(y, infer = "MCMC", D = 1.702, mu_th = 0, sigma_th = 1,
  mu_b = 0, sigma_b = 3, location_a = 0, scale_a = 3,
  minimam_response = 1, ...)
}
\arguments{
\item{y}{integer data.frame contains only item response. Do NOT contain subject ID and group index.}

\item{infer}{a method of inference. "MCMC" or "VB"}

\item{D}{scale constant}

\item{mu_th, sigma_th}{a hyper prameter of theta. mu and sigma of normal distribution.}

\item{mu_b, sigma_b}{a hyper parameter of difficulty. mu and sigma of normal distribution}

\item{location_a, scale_a}{a hyper parameter of discrimination. location and scale of Cauchy distribution.}

\item{...}{Arguments passe to \code{\link[rstan]{sampling}} or \code{\link[rstan]{vb}}.}
}
\description{
Bayesian Graded Response modelng fitting
}
