% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtstan-package.R
\docType{package}
\name{irtstan-package}
\alias{irtstan-package}
\alias{irtstan}
\title{The 'irtstan' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
