% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girtmodel_stan.R
\name{estGip_stan}
\alias{estGip_stan}
\title{Bayesian GIRT model fitting}
\usage{
estGip_stan(y, ...)
}
\arguments{
\item{y}{integer matrix contains only item response. Do NOT contain subject ID and group index.}

\item{...}{Arguments passe to \code{\link{rstan::sampling}}}
}
\value{
stanfit class file returned by \code{\link{rstan::sampling}}
}
\description{
Bayesian GIRT model fitting
}
