% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_wrapper.R
\name{estip_stan}
\alias{estip_stan}
\title{Bayesian 2PL modelng fitting}
\usage{
estip_stan(y, model = "2PL", infer = "MCMC", D = 1.702, mu_th = 0,
  sigma_th = 1, mu_b = 0, sigma_b = 3, location_a = 0,
  scale_a = 3, alpha_c = 1, beta_c = 3, max_scale = 5,
  min_scale = -5, ...)
}
\arguments{
\item{y}{integer matrix contains only item response. Do NOT contain subject ID and group index.}

\item{model}{irt model type. 1~3PLM.}

\item{infer}{a method of inference. "MCMC" or "VB"}

\item{D}{scale constant}

\item{mu_th, sigma_th}{a hyper prameter of theta. mu and sigma of normal distribution.}

\item{mu_b, sigma_b}{a hyper parameter of difficulty. mu and sigma of normal distribution}

\item{location_a, scale_a}{a hyper parameter of discrimination. location and scale of Cauchy distribution.}

\item{max_scale, min_scale}{maximum or minimum value in theta and b parameter scale.}

\item{...}{Arguments passe to \code{\link[rstan]{sampling}} or \code{\link[rstan]{vb}}.}
}
\description{
Bayesian 2PL modelng fitting
}
